refactor: centralize log list logic
Move CT log list management to a dedicated module. The associated test file is also renamed.
