+++ internal/certificatetransparency/ct-watcher.go
@@ -564,116 +564,6 @@ func certHandler(entryChan chan models.Entry) {
 	}
 }
 
-// getAllLogs returns a list of all CT logs - those from the Google list, if not disabled -
-// and additional logs provided via the config.
-func getAllLogs(logListFetcher LogListFetcher) (loglist3.LogList, error) {
-	var allLogs loglist3.LogList
-
-	// Ability to disable default logs, if the user only wants to monitor custom logs.
-	if !config.AppConfig.General.DisableDefaultLogs {
-		var err error
-
-		allLogs, err = logListFetcher.Fetch()
-		if err != nil {
-			log.Printf("Error fetching log list: %s\n", err)
-			return loglist3.LogList{}, fmt.Errorf("failed to fetch log list: %w", err)
-		}
-	}
-
-	// Add additional logs provided via the config, if any.
-	for _, additionalLog := range config.AppConfig.General.AdditionalLogs {
-		customLog := loglist3.Log{
-			URL:         additionalLog.URL,
-			Description: additionalLog.Description,
-		}
-
-		operatorFound := false
-
-		for _, operator := range allLogs.Operators {
-			// Only compare logs with the same operator
-			if operator.Name != additionalLog.Operator {
-				continue
-			}
-
-			operatorFound = true
-			logFound := false
-
-			// Check if user provided log is already in our loglist
-			for _, ctlog := range operator.Logs {
-				if ctlog.URL == additionalLog.URL {
-					// Log already exists, skip it.
-					logFound = true
-					break
-				}
-			}
-
-			if !logFound {
-				// This works, since allLogs.Operators is a slice of pointers.
-				operator.Logs = append(operator.Logs, &customLog)
-			}
-
-			break
-		}
-
-		if !operatorFound {
-			newOperator := loglist3.Operator{
-				Name: additionalLog.Operator,
-				Logs: []*loglist3.Log{&customLog},
-			}
-			allLogs.Operators = append(allLogs.Operators, &newOperator)
-		}
-	}
-
-	for _, additionalLog := range config.AppConfig.General.AdditionalTiledLogs {
-		customLog := loglist3.TiledLog{
-			MonitoringURL: additionalLog.URL,
-			Description:   additionalLog.Description,
-		}
-
-		operatorFound := false
-
-		for _, operator := range allLogs.Operators {
-			if operator.Name == additionalLog.Operator {
-				operatorFound = true
-				logFound := false
-
-				for _, tl := range operator.TiledLogs {
-					if tl.MonitoringURL == additionalLog.URL {
-						// Log already exists, skip it.
-						logFound = true
-						break
-					}
-				}
-
-				if !logFound {
-					// This works, since allLogs.Operators is a slice of pointers.
-					operator.TiledLogs = append(operator.TiledLogs, &customLog)
-				}
-
-				break
-			}
-		}
-
-		if !operatorFound {
-			newOperator := loglist3.Operator{
-				Name:      additionalLog.Operator,
-				TiledLogs: []*loglist3.TiledLog{&customLog},
-			}
-			allLogs.Operators = append(allLogs.Operators, &newOperator)
-		}
-	}
-
-	return allLogs, nil
-}
-
-func normalizeCtlogURL(input string) string {
-	input = strings.TrimPrefix(input, "https://")
-	input = strings.TrimPrefix(input, "http://")
-	input = strings.TrimSuffix(input, "/")
-
-	return input
-}
-
 // newHTTPClient creates a new http.Client with reasonable timeouts and connection settings for interacting with CT logs.
 func newHTTPClient() *http.Client {
 	httpClient := http.Client{
